Move tutorial starter code from week2 folder to week3 folder, since we moved the tutorial itself.
