+++ tutorials/week3/dictionary.cpp
@@ -3,7 +3,7 @@
 #include <string>
 #include <vector>
 
-#include "tutorials/week2/dictionary_library.h"
+#include "tutorials/week3/dictionary_library.h"
 
 int main() {
   std::ifstream word_list_file = std::ifstream{"data/words.txt"};
